Create separate packages for tests which switch the Arwen version
